+++ System.Net.Mqtt.Server/MqttServerLoggingExtensions.cs
@@ -51,8 +51,8 @@ internal static partial class MqttServerLoggingExtensions
     public static partial void LogSessionTerminatedGracefully(this ILogger logger, MqttServerSession session);
 
     [LoggerMessage(16, Debug, "Incoming message from '{clientId}': Topic = '{topic}', Size = {size}, QoS = {qos}, Retain = {retain}", EventName = "IncomingMessage", SkipEnabledCheck = true)]
-    public static partial void LogIncomingMessage(this ILogger logger, string clientId, string topic, int size, byte qos, bool retain);
+    public static partial void LogIncomingMessage(this ILogger logger, string clientId, string topic, int size, int qos, bool retain);
 
     [LoggerMessage(17, Debug, "Outgoing message for '{clientId}': Topic = '{topic}', Size = {size}, QoS = {qos}, Retain = {retain}", EventName = "OutgoingMessage", SkipEnabledCheck = true)]
-    public static partial void LogOutgoingMessage(this ILogger logger, string clientId, string topic, int size, byte qos, bool retain);
+    public static partial void LogOutgoingMessage(this ILogger logger, string clientId, string topic, int size, int qos, bool retain);
 }
